% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{culturalEvolution-package}
\alias{culturalEvolution}
\alias{culturalEvolution-package}
\title{culturalEvolution: Simple Workflow for Cultural Evolution Models}
\description{
This package provides convenient functions for working with models of cultural evolution and/or diffusion processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acastroaraujo/culturalEvolution}
  \item Report bugs at \url{https://github.com/acastroaraujo/culturalEvolution/issues}
}

}
\keyword{internal}
