% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{multi_par_call}
\alias{multi_par_call}
\title{Run Recursive Model with Multiple Parameters}
\usage{
multi_par_call(grid, model, q_init = 0, tn = 20)
}
\arguments{
\item{grid}{a data frame of parameter values}

\item{model}{a function of class "recursion."}

\item{q_init}{the initial value for "q" (default is zero).}

\item{tn}{the number of time periods after q_0}
}
\value{
a data frame with parameters t, q, .id, and whatever parameters where used
}
\description{
Run Recursive Model with Multiple Parameters
}
\examples{

env_learn <- recursion(
  expr = P1 + L*q
)

env_learn

grid <- expand.grid(P1 = seq(0, 0.5, by  = 0.1), L = seq(0.5, 1, by = 0.1))

multi_par_call(grid, env_learn)

}
