% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{recursion}
\alias{recursion}
\title{Create Recursive Models}
\usage{
recursion(var, expr)
}
\arguments{
\item{var}{The state variable or "trait" proportion (e.g., q). You are not
allowed to choose "t", which is reserved for keeping track of time.}

\item{expr}{A recursive mathematical expression.}
}
\value{
A function of class "rfun", with the following arguments:

\describe{
  \item{params}{A list or data frame of parameter values.}
  \item{init}{The initial value for the state variable (defaults to zero).}
  \item{tn}{The number of time periods.}
}
}
\description{
Create Recursive Models
}
\examples{
biased_transmission <- recursion(var = q, expr = q + q*(1 - q)*B)
biased_transmission

}
