% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{recursion}
\alias{recursion}
\title{Create Recursive Models}
\usage{
recursion(expr)
}
\arguments{
\item{expr}{a mathematical expression.}
}
\value{
a function of class "rfun."
}
\description{
Create Recursive Models
}
\examples{
biased_transmission <- recursion(expr = q + q*(1 - q)*B)
biased_transmission

}
